# examples/axum-example/src/observations/domain.rs
use jiff::Timestamp;
use serde::{Deserialize, Serialize};
use utoipa::ToSchema;
use uuid::Uuid;

#[derive(
    Debug,
    Clone,
    Copy,
    PartialEq,
    Eq,
    PartialOrd,
    Ord,
    Hash,
    Serialize,
    Deserialize,
    ToSchema,
    derive_more::Display,
)]
#[display("{_0}")]
pub struct ObservationId(Uuid);

impl ObservationId {
    #[must_use]
    pub fn generate() -> Self {
        Self(Uuid::now_v7())
    }
}

#[derive(Debug, Clone, PartialEq, Serialize, Deserialize, ToSchema)]
pub struct PartialObservation {
    pub name: String,
    pub position: LngLat,
    pub color: Option<String>,
    pub notes: Option<String>,
}

#[derive(Debug, Clone, PartialEq, Serialize, Deserialize, ToSchema)]
pub struct Observation {
    pub id: ObservationId,
    #[schema(value_type = String, example = "2024-01-01T00:00:00Z")]
    pub created_at: Timestamp,

    #[serde(flatten)]
    pub data: PartialObservation,
}

#[derive(Debug, Clone, Copy, PartialEq, PartialOrd, Serialize, Deserialize, ToSchema)]
pub struct LngLat {
    pub lng: f64,
    pub lat: f64,
}

#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize, ToSchema)]
pub struct PatchObservation {
    #[serde(skip_serializing_if = "Option::is_none")]
    pub name: Option<String>,
    #[serde(skip_serializing_if = "Option::is_none")]
    pub position: Option<LngLat>,
    #[serde(skip_serializing_if = "Option::is_none")]
    pub color: Option<String>,
    #[serde(skip_serializing_if = "Option::is_none")]
    pub notes: Option<String>,
}

#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize, ToSchema)]
pub struct SearchObservations {
    #[serde(skip_serializing_if = "Option::is_none")]
    pub name: Option<String>,
    #[serde(skip_serializing_if = "Option::is_none")]
    pub color: Option<String>,
}

impl SearchObservations {
    #[must_use]
    pub fn matches(&self, observation: &PartialObservation) -> bool {
        let name_matches = self.name.as_deref().is_none_or(|name| {
            observation
                .name
                .to_lowercase()
                .contains(&name.to_lowercase())
        });
        let color_matches = self
            .color
            .as_deref()
            .is_none_or(|color| observation.color.as_deref() == Some(color));
        name_matches && color_matches
    }
}

#[derive(Debug, Clone, PartialEq, Serialize, Deserialize, ToSchema)]
pub struct ImportResponse {
    pub imported: usize,
    pub error_count: usize,
    pub bytes_processed: usize,
    pub created_ids: Vec<ObservationId>,
    #[serde(skip_serializing_if = "Option::is_none")]
    pub errors: Option<Vec<String>>,
}

#[derive(Debug, Clone, PartialEq, Serialize, Deserialize, ToSchema)]
pub struct UploadResponse {
    pub uploaded: usize,
    pub error_count: usize,
    pub created_ids: Vec<ObservationId>,
    #[serde(skip_serializing_if = "Option::is_none")]
    pub errors: Option<Vec<String>>,
}